refactor(`mock-edupass`): inline accounts and interaction modules into their consumers
