--- apps/mock-edupass/src/accounts.ts
@@ -1,5 +1,0 @@
-export const accounts = [
-  { sub: 'teacher-1', email: 'jane.doe@example.com', name: 'Jane Doe' },
-  { sub: 'teacher-2', email: 'john.smith@example.com', name: 'John Smith' },
-  { sub: 'teacher-3', email: 'no-name@example.com' },
-];
