+++ apps/mock-edupass/src/provider.ts
@@ -1,6 +1,10 @@
 import Provider from 'oidc-provider';
 
-import { accounts } from './accounts.ts';
+export const accounts = [
+  { sub: 'teacher-1', email: 'jane.doe@example.com', name: 'Jane Doe' },
+  { sub: 'teacher-2', email: 'john.smith@example.com', name: 'John Smith' },
+  { sub: 'teacher-3', email: 'no-name@example.com' },
+];
 
 export function createProvider(port: number): Provider {
   const issuer = `http://localhost:${port}`;
